move damage effect to different package
